requirements.txt
# AI Daily Analyst - Requirements

# Web Search
tavily-python>=0.7.0

# HTTP requests
requests>=2.31.0

# WeChat SDK (可选)
# wechatpy>=1.8.0

# 日期处理
python-dateutil>=2.8.0
